pyproject.toml
[tool.poetry]
name = "диплом"
version = "0.1.0"
description = ""
authors = ["Павел Садальский <pavel.sadalsky@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
loguru = "^0.6.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
